backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-chess==1.999
stockfish==3.28.0
pydantic==2.10.3


